refactor(scheduler): move the PD-only states into pd_states.h
RemotePrefilling and PrefillAwaitingResult are role-exclusive: the
first exists only while a D engine holds destination pages for the
peer's prefill, the second only while a P engine holds a finished
prompt for the bootstrap token its remote decode needs. A fused engine
can never enter either, exactly like Bootstrapping, which already
lives in pd_states.h -- so colocate them and leave forward_states.h
holding just the universal lifecycle.

pd_states.h now includes forward_states.h for the state bases, and
forward_events.h includes pd_states.h: the D and P grammars' schedule
events construct these states, so the dependency existed all along --
the move just makes it visible at the file boundary.
